+++ application/src/main/resources/static/js/react-client/src/pages/userhome-screen/index.js
@@ -5,9 +5,14 @@ import UserInfo from '../../userinfo';
 import prfBg from '../../assets/profile-screen/prfBackground.jpg';
 import style from './style.module.scss';
 function UserScreen() {
+  let display = true; 
+  if(localStorage.getItem("token")==""){
+  window.location.replace("/login"); display = false; }
+  
   return (
-
+   
     <div className={style.MainProfileDiv}>
+      { display==false?<></>:
         <div className={style.profileContainer}>
         <div className={style.topPortion}>
         <div className={style.userProfileBgImage}>
@@ -30,8 +35,10 @@ function UserScreen() {
         </div>
       </div>
         </div>
+   } 
     </div>
   );
+  
 }
 
 export default UserScreen;
