+++ application/src/main/resources/static/js/react-client/src/pages/userhome-screen/style.module.scss
@@ -63,7 +63,7 @@
         display: flex;
         justify-content: center;
         flex-direction: row;
-        width: 90vw;
+        width: 100%;
         margin-top: 5rem;
         background: rgb(255, 255, 255);
     }
